refactor: app 모듈 구조 변경(#15)
